Server/requirements.txt
qbittorrent-api>=0.3.2
python-qbittorrent>=0.3.1
PlexAPI>=3.1.0
Twisted>=19.2.1
psutil
